django/requirements.txt
boto3==1.9.160
botocore==1.12.160
certifi==2019.3.9
chardet==3.0.4
Django==2.1.0
django-extensions==2.1.7
get-docker-secret
django-storages==1.7.1
docutils==0.14
idna==2.8
jmespath==0.9.4
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
s3transfer==0.2.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
channels
asgi_redis
channels_redis
celery
psycopg2
graphene-django
graphene-django-extras
#rest_framework_jwt
django-channels-graphql-ws
rx==1.6.0
djangorestframework
djangorestframework-jwt
markdown
django-filter
django-webpack-loader
ptvsd
aiohttp
dataclasses
attrdict